[hiptest] Add necessary cast

+++ src/hiptest/test/world.cu
@@ -5,7 +5,7 @@
 
 #include "CUDACore/cudaCheck.h"
 
-__global__ void print() { printf("GPU thread %d\n", threadIdx.x); }
+__global__ void print() { printf("GPU thread %d\n", static_cast<int>(threadIdx.x)); }
 
 int main() {
   std::cout << "World from" << std::endl;
